添加注释，修改部分多线下载逻辑

+++ download/src/main/java/com/gioneco/download/bean/DownloadInfo.kt
@@ -1,13 +1,16 @@
 package com.gioneco.download.bean
 
 /**
+ * 下载信息
+ * 注意如果字段有所变更，需更新数据库[com.gioneco.download.db.DBHelper]字段
+ *
  *Created by zsq
  *on 2021-01-08
  */
 data class DownloadInfo(
     val threadId: Int,// 下载器id
-    val startPos: Int,// 开始点
-    val endPos: Int,// 结束点
-    val completeSize: Int,// 完成度
+    val startPos: Long,// 开始点
+    val endPos: Long,// 结束点
+    val completeSize: Long,// 完成度
     val url: String// 下载器网络标识
 )
